feat: migrate profile info aggregate to nestjs
Ports GET /profile/info (the profile page aggregate) to
GET /api/v2/profile/info (getFullProfileInfo operation, DefaultGuard).

The whole legacy implementation is ported preserving business logic:
- permission matrix (defineRole/getPermissions/
  getProfilePermissionsSettings) moved verbatim into
  nestjs/src/profile/profile-info/permissions.ts together with the
  configurable-permissions merge and the relations-role resolution
  (student/mentor/interviewer/checker relations, course manager and
  supervisor checks, mentor registry based course discovery);
- data queries (user info with permission-driven selects, mentor
  stats, public feedback, stage interview feedback incl. legacy
  feedback template parsing via the shared getInterviewRatings,
  student stats with rank/certificate/task aggregation) ported
  verbatim into ProfileInfoService on top of the shared DataSource;
- the orchestrator keeps the same visibility gating: 403 for
  non-owners without isProfileVisible, optional sections fetched only
  when the respective permission allows, own stage interview feedback
  is never returned.

Equivalence verification: the existing 463-line legacy permissions
test-suite (20 tests) is mirrored 1:1 in nestjs (permissions.spec.ts)
and passes against the ported matrix; the orchestrator is covered by
identical scenario tests on both sides (owner shape, stranger 403,
admin full access). Legacy scaffolding tests were removed with the
legacy code; the permission tests now live in nestjs.

Frontend: UserService.getProfileInfo switched to the generated
ProfileApi.getFullProfileInfo.
