+++ client/src/api/api.ts
@@ -5330,6 +5330,61 @@ export interface ProfileDto {
 /**
  * 
  * @export
+ * @interface ProfileInfoExtendedDto
+ */
+export interface ProfileInfoExtendedDto {
+    /**
+     * 
+     * @type {object}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'permissionsSettings'?: object;
+    /**
+     * 
+     * @type {object}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'generalInfo'?: object;
+    /**
+     * 
+     * @type {object}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'contacts'?: object;
+    /**
+     * 
+     * @type {object}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'discord'?: object | null;
+    /**
+     * 
+     * @type {Array<object>}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'mentorStats'?: Array<object>;
+    /**
+     * 
+     * @type {Array<object>}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'publicFeedback'?: Array<object>;
+    /**
+     * 
+     * @type {Array<object>}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'stageInterviewFeedback'?: Array<object>;
+    /**
+     * 
+     * @type {Array<object>}
+     * @memberof ProfileInfoExtendedDto
+     */
+    'studentStats'?: Array<object>;
+}
+/**
+ * 
+ * @export
  * @interface PromptDto
  */
 export interface PromptDto {
@@ -9534,6 +9589,35 @@ export const CertificateApiAxiosParamCreator = function (configuration?: Configu
         },
         /**
          * 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        getCertificateTemplates: async (options: AxiosRequestConfig = {}): Promise<RequestArgs> => {
+            const localVarPath = `/certificate/templates`;
+            // use dummy base URL string because the URL constructor only accepts absolute URLs.
+            const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL);
+            let baseOptions;
+            if (configuration) {
+                baseOptions = configuration.baseOptions;
+            }
+
+            const localVarRequestOptions = { method: 'GET', ...baseOptions, ...options};
+            const localVarHeaderParameter = {} as any;
+            const localVarQueryParameter = {} as any;
+
+
+    
+            setSearchParams(localVarUrlObj, localVarQueryParameter);
+            let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
+            localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
+
+            return {
+                url: toPathString(localVarUrlObj),
+                options: localVarRequestOptions,
+            };
+        },
+        /**
+         * 
          * @param {number} studentId 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -9622,6 +9706,15 @@ export const CertificateApiFp = function(configuration?: Configuration) {
         },
         /**
          * 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        async getCertificateTemplates(options?: AxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<void>> {
+            const localVarAxiosArgs = await localVarAxiosParamCreator.getCertificateTemplates(options);
+            return createRequestFunction(localVarAxiosArgs, globalAxios, BASE_PATH, configuration);
+        },
+        /**
+         * 
          * @param {number} studentId 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -9661,6 +9754,14 @@ export const CertificateApiFactory = function (configuration?: Configuration, ba
         },
         /**
          * 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        getCertificateTemplates(options?: any): AxiosPromise<void> {
+            return localVarFp.getCertificateTemplates(options).then((request) => request(axios, basePath));
+        },
+        /**
+         * 
          * @param {number} studentId 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -9696,6 +9797,16 @@ export class CertificateApi extends BaseAPI {
      */
     public getCertificate(publicId: string, options?: AxiosRequestConfig) {
         return CertificateApiFp(this.configuration).getCertificate(publicId, options).then((request) => request(this.axios, this.basePath));
+    }
+
+    /**
+     * 
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     * @memberof CertificateApi
+     */
+    public getCertificateTemplates(options?: AxiosRequestConfig) {
+        return CertificateApiFp(this.configuration).getCertificateTemplates(options).then((request) => request(this.axios, this.basePath));
     }
 
     /**
@@ -17646,6 +17757,40 @@ export const ProfileApiAxiosParamCreator = function (configuration?: Configurati
         },
         /**
          * 
+         * @param {string} [githubId] 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        getFullProfileInfo: async (githubId?: string, options: AxiosRequestConfig = {}): Promise<RequestArgs> => {
+            const localVarPath = `/profile/info`;
+            // use dummy base URL string because the URL constructor only accepts absolute URLs.
+            const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL);
+            let baseOptions;
+            if (configuration) {
+                baseOptions = configuration.baseOptions;
+            }
+
+            const localVarRequestOptions = { method: 'GET', ...baseOptions, ...options};
+            const localVarHeaderParameter = {} as any;
+            const localVarQueryParameter = {} as any;
+
+            if (githubId !== undefined) {
+                localVarQueryParameter['githubId'] = githubId;
+            }
+
+
+    
+            setSearchParams(localVarUrlObj, localVarQueryParameter);
+            let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
+            localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
+
+            return {
+                url: toPathString(localVarUrlObj),
+                options: localVarRequestOptions,
+            };
+        },
+        /**
+         * 
          * @param {string} username 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -17878,6 +18023,16 @@ export const ProfileApiFp = function(configuration?: Configuration) {
         },
         /**
          * 
+         * @param {string} [githubId] 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        async getFullProfileInfo(githubId?: string, options?: AxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<ProfileInfoExtendedDto>> {
+            const localVarAxiosArgs = await localVarAxiosParamCreator.getFullProfileInfo(githubId, options);
+            return createRequestFunction(localVarAxiosArgs, globalAxios, BASE_PATH, configuration);
+        },
+        /**
+         * 
          * @param {string} username 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -17966,6 +18121,15 @@ export const ProfileApiFactory = function (configuration?: Configuration, basePa
         },
         /**
          * 
+         * @param {string} [githubId] 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        getFullProfileInfo(githubId?: string, options?: any): AxiosPromise<ProfileInfoExtendedDto> {
+            return localVarFp.getFullProfileInfo(githubId, options).then((request) => request(axios, basePath));
+        },
+        /**
+         * 
          * @param {string} username 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -18048,6 +18212,17 @@ export class ProfileApi extends BaseAPI {
      */
     public getEndorsementData(username: string, options?: AxiosRequestConfig) {
         return ProfileApiFp(this.configuration).getEndorsementData(username, options).then((request) => request(this.axios, this.basePath));
+    }
+
+    /**
+     * 
+     * @param {string} [githubId] 
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     * @memberof ProfileApi
+     */
+    public getFullProfileInfo(githubId?: string, options?: AxiosRequestConfig) {
+        return ProfileApiFp(this.configuration).getFullProfileInfo(githubId, options).then((request) => request(this.axios, this.basePath));
     }
 
     /**
