+++ nestjs/src/profile/dto/index.ts
@@ -1,3 +1,4 @@
 export * from './profile-course.dto';
 export * from './update-user.dto';
 export * from './update-profile.dto';
+export * from './profile-info-extended.dto';
