+++ server/src/routes/profile/index.ts
@@ -1,27 +1,11 @@
 import Router from '@koa/router';
 import { ILogger } from '../../logger';
 import { guard } from '../guards';
-import { getProfileInfo } from './info';
 import { getMyProfile, updateMyProfile } from './me';
 
 export function profileRoute(logger: ILogger) {
   const router = new Router<any, any>({ prefix: '/profile' });
 
-  /**
-   * @swagger
-   *
-   * /profile:
-   *   get:
-   *      description: get user profile
-   *      security:
-   *        - cookieAuth: []
-   *      produces:
-   *        - application/json
-   *      responses:
-   *        200:
-   *          description: profile
-   */
-  router.get('/info', guard, getProfileInfo(logger));
 
   /**
    * @swagger
